% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{assessMetrics}
\alias{assessMetrics}
\title{Provides broad-stroke data quality assessment}
\usage{
assessMetrics(ploidytest)
}
\arguments{
\item{ploidytest}{ploidytest object}
}
\value{
Vector of quality calls
}
\description{
assessMetrics looks at the residual score, the number of genes considered, and the fraction of 0's
in the considered genes and provides a simple quality assessment on each. This assumes default parameter choice.
}
