% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calcAneu}
\alias{calcAneu}
\title{Calculate aneuploidy scores for one cell group}
\usage{
calcAneu(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
ploidytest object with aneuploidy results
}
\description{
This is the guts of the aneuploidy calling method, which is called by doAneu().
You probably don't want to call this yourself
}
