% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{doAneu}
\alias{doAneu}
\title{Perform Aneuploidy Assessment}
\usage{
doAneu(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Ploidytest object with results
}
\description{
doAneu performs the aneuploidy assessment.
}
