% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getCellNames}
\alias{getCellNames}
\title{Get cell names from ploidytest object}
\usage{
getCellNames(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Cell name vector
}
\description{
Get cell names from ploidytest object
}
