% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getCHR}
\alias{getCHR}
\title{Get chromosome locations for each gene from ploidytest object}
\usage{
getCHR(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Chromosome vector for each gene
}
\description{
Get chromosome locations for each gene from ploidytest object
}
