% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getCounts}
\alias{getCounts}
\title{Get counts matrix from ploidytest object}
\usage{
getCounts(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Counts matrix
}
\description{
Get counts matrix from ploidytest object
}
