% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getCPM}
\alias{getCPM}
\title{Get CPM matrix from ploidytest object}
\usage{
getCPM(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
CPM matrix
}
\description{
Get CPM matrix from ploidytest object
}
