% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getFN}
\alias{getFN}
\title{Get call scores for False negative aneuploidy calls}
\usage{
getFN(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object with known aneuploidies}
}
\value{
Score data frame for false negative calls only
}
\description{
Get call scores for False negative aneuploidy calls
}
