% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getFP}
\alias{getFP}
\title{Get call scores for False positive aneuploidy calls}
\usage{
getFP(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object with known aneuploidies}
}
\value{
Score data frame for false positive calls only
}
\description{
Get call scores for False positive aneuploidy calls
}
