% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getGenes}
\alias{getGenes}
\title{Get gene names from ploidytest object}
\usage{
getGenes(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Gene name vector
}
\description{
Get gene names from ploidytest object
}
