% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getGeneTable}
\alias{getGeneTable}
\title{Get a gene-CHR map}
\usage{
getGeneTable(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
data frame with gene and chr columns
}
\description{
Get a gene-CHR map
}
