% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getGroups}
\alias{getGroups}
\title{Get cell groups from ploidytest object}
\usage{
getGroups(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Cell group vector
}
\description{
Get cell groups from ploidytest object
}
