% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getHits}
\alias{getHits}
\title{Get prediction output from ploidytest object for predicted aneuploidies only}
\usage{
getHits(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Data frame of prediction outputs for aneuploid-called chromosomes. Contains:
chr: chromosome
cell: cell
z: z-score for p-value calculation
score: s_{ij}
p: raw p-value
p.adj: p-value FDR-adjusted 
monosomy: TRUE if z<0 i.e. reduced expression
}
\description{
Get prediction output from ploidytest object for predicted aneuploidies only
}
