% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getKnownAneu}
\alias{getKnownAneu}
\title{Get known aneuploidies from ploidytest object}
\usage{
getKnownAneu(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Data frame of known aneuploidies
}
\description{
Get known aneuploidies from ploidytest object
}
