% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getMaxA}
\alias{getMaxA}
\title{Gets the maximum value of a_{gij} that each considered gene contributes}
\usage{
getMaxA(ploidytest)
}
\arguments{
\item{ploidytest}{ploidytest object}
}
\value{
Vector of all genes considered for analysis in any batch, named with the gene ID, containing the maximum
a_{gij} score observed
}
\description{
Gets the maximum value of a_{gij} that each considered gene contributes
}
