% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getMeanVar}
\alias{getMeanVar}
\title{Get mean-variance relationship of genes}
\usage{
getMeanVar(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Data frame of mean (gene mean CPM) and sd (gene CPM sd)
}
\description{
Function gets the mean and variance of expression for each gene's CPM that would be used in aneuploidy
assessment
}
