% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getMetrics}
\alias{getMetrics}
\title{Get data metrics to estimate performance}
\usage{
getMetrics(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
list of data metrics:
\itemize{
\item "ngenes": number of genes that qualify for aneuploidy assessment (median CPM>50 only)
\item "zeros": fraction of 0 counts for each gene that qualifies for aneuploidy assessment (median CPM>50 only) 
\item "residual": residual score vs. G&T-seq 8-cell stage embryos (see supplemental document of paper)
}
}
\description{
Get data metrics to estimate performance
}
