% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getParam}
\alias{getParam}
\title{Get parameter for aneuploidy calling}
\usage{
getParam(ploidytest, param_name)
}
\arguments{
\item{ploidytest}{Ploidytest object}

\item{param_name}{Name of parameter (see ?getParams or ?setParams)}
}
\value{
named parameter value
}
\description{
Get parameter for aneuploidy calling
}
