% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getParams}
\alias{getParams}
\title{Get parameters for aneuploidy calling}
\usage{
getParams(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Parameter list. Contains:
\itemize{
\item "p.thresh": post FDR correction p-value for aneuploidy calling (default 0.1)
\item "min.deviation": minimum deviation of s_{ij} from 1 for aneuploidy calling (default 0.2)
\item "center.cells": logical value to allow or prevent within-cell s_{ij} centering (default TRUE)
\item "min.median": minimum CPM median value for gene retention for analysis (default 50)
\item "extreme.gene.thresh": maximum a_{gij} that one gene may show before removal as outlier (default Inf)
}
}
\description{
Get parameters for aneuploidy calling
}
