% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getScores}
\alias{getScores}
\title{Get prediction output from ploidytest object}
\usage{
getScores(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Data frame of prediction outputs for all chromosomes. Contains:
\itemize{
\item chr: chromosome
\item cell: cell
\item z: z-score for p-value calculation
\item score: s_{ij}
\item p: raw p-value
\item p.adj: p-value FDR-adjusted 
\item monosomy: TRUE if z<0 i.e. reduced expression - *not* an indicator of aneuploidy
}
}
\description{
Get prediction output from ploidytest object
}
