% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getSijMat}
\alias{getSijMat}
\title{Get matrix of sij}
\usage{
getSijMat(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Sij matrix, rows chromosomes, columns cells
}
\description{
Get matrix of sij
}
