% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{getTrueScores}
\alias{getTrueScores}
\title{Get call scores for all true aneuploidies}
\usage{
getTrueScores(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object with known aneuploidies}
}
\value{
Score data frame for true aneuploidies
}
\description{
Get call scores for all true aneuploidies
}
