% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{makeAneu}
\alias{makeAneu}
\title{Aneuploidy Constructor function}
\usage{
makeAneu(counts = NULL, genes = NULL, chrs = NULL, cellNames = NULL,
  cellGroups = NULL, params = NULL)
}
\arguments{
\item{counts}{Raw counts matrix (i.e. integers)}

\item{genes}{Genes for the rows of the counts matrix}

\item{chrs}{Chromosome number for each of the genes}

\item{cellNames}{Names for each of the columns of the counts matrix}

\item{cellGroups}{Cell groups for use in analysis (i.e. different cell types)}
}
\value{
Ploidytest object
}
\description{
This function is the constructor for the ploidytest object
}
