% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot a PCA of log-CPM counts}
\usage{
plotPCA(ploidytest, cols = NULL)
}
\arguments{
\item{ploidytest}{Ploidytest object}

\item{cols}{Colour labels e.g. cell line or embryo number}
}
\value{
invisible(0)
}
\description{
Plot a PCA of log-CPM counts
}
