% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{presentFN}
\alias{presentFN}
\title{Present False Negative Results}
\usage{
presentFN(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
data frame of false negative calls
}
\description{
Function returns a nicely formatted data frame of false negative calls, when aneuploidies are known
}
