% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{presentKnown}
\alias{presentKnown}
\title{Present Known Aneuploidy Scores}
\usage{
presentKnown(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
data frame of true aneuploidy calls
}
\description{
Function returns a nicely formatted data frame of scores for true aneuploidies
}
