% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sampleCells}
\alias{sampleCells}
\title{Randomly downsample cells}
\usage{
sampleCells(ploidytest, n)
}
\arguments{
\item{ploidytest}{Ploidytest object}

\item{n}{Number of cells to sample}
}
\value{
Ploidytest object with fewer cells
}
\description{
Randomly downsample cells
}
