% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{setKnownAneu}
\alias{setKnownAneu}
\title{Add known aneuploidy data to ploidytest object}
\usage{
setKnownAneu(ploidytest, aneu_df)
}
\arguments{
\item{ploidytest}{Ploidytest object}

\item{aneu_df}{Data frame of aneuploidies. Requires columns:
\itemize{
\item "chr": Chromosome
\item "cell": Cell name
\item "monosomy": TRUE if monosomy, FALSE if trisomy
}}
}
\value{
Ploidytest object with known aneuploidies
}
\description{
Add known aneuploidy data to ploidytest object
}
