% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{setParam}
\alias{setParam}
\title{Set parameters for aneuploidy calling}
\usage{
setParam(ploidytest, param_name, param_value, print = FALSE)
}
\arguments{
\item{ploidytest}{Ploidytest object}

\item{param_name}{Name of parameter. May be any of:
\itemize{
\item "p.thresh": post FDR correction p-value for aneuploidy calling (default 0.1)
\item "min.deviation": minimum deviation of s_{ij} from 1 for aneuploidy calling (default 0.2)
\item "center.cells": logical value to allow or prevent within-cell s_{ij} centering (default TRUE)
\item "min.median": minimum CPM median value for gene retention for analysis (default 50)
\item "extreme.gene.thresh": maximum a_{gij} that one gene may show before removal as outlier (default Inf)
}}

\item{param_value}{New parameter value}

\item{print}{Prints the changes to the numbers of genes considered if you change extreme.gene.thresh if TRUE}
}
\value{
Ploidytest object with revised parameter value
}
\description{
Set parameters for aneuploidy calling
}
