% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{splitCellsByGroup}
\alias{splitCellsByGroup}
\title{Separate cell analysis groups}
\usage{
splitCellsByGroup(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
List of ploidytest objects, names according to the cellGroups category represented.
}
\description{
This function splits your cells into the cellGroups specified in the constructor.
}
