% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subsetGenes}
\alias{subsetGenes}
\title{Subset genes in ploidytest object}
\usage{
subsetGenes(ploidytest, logical_keep)
}
\arguments{
\item{ploidytest}{Ploidytest object}

\item{logical_keep}{Logical vector of genes to retain (TRUE) and drop (FALSE)}
}
\value{
Ploidytest object with (probably) fewer genes
}
\description{
Subset genes in ploidytest object
}
