% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{testPerformance}
\alias{testPerformance}
\title{Test the method's performance where aneuploidy is known}
\usage{
testPerformance(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Vector of performance metrics
}
\description{
Test the method's performance where aneuploidy is known
}
