% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{trimGenes}
\alias{trimGenes}
\title{Exclude lowly expressed and DE genes (if specified)}
\usage{
trimGenes(ploidytest)
}
\arguments{
\item{ploidytest}{Ploidytest object}
}
\value{
Ploidytest object with fewer genes
}
\description{
This function excludes genes from analysis and is called during doAneu()
}
